refactor: 기사 목록 컴포넌트 이동
